services/signal-gateway/requirements.txt
-e ../../packages/common-schemas/python
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
httpx==0.27.2
python-dotenv
psycopg[binary]
